Began restructuring the help panel.
Increased size of the window along with the help panel width.
Added a new function to load help panel content.
Added a new field in the sections config that defines what content should be put in the help panel.
Current help panel types are helpText and helpInsert. The helpInsert adds a button that inserts text into the configured input box.