Finished new help panel logic and config. Added save dialog.
Removed infoBarText.json as it is no longer needed.
Moved all infoBarText.json content over to sections.json with the new help panel configuration structure.
Finished formatting the help panel for readability.
Added an auto font-size scaling function for when the user resizes the window horizontally.
Added a save dialog so the user can choose where to put and name the output file.